requirements.txt
# Core Framework
flask>=2.3.0
flask-cors>=4.0.0
Werkzeug>=2.3.0

# Text Extraction
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
docx2txt>=0.8
PyPDF2>=3.0.0

# NLP & ML
sentence-transformers>=2.2.0
transformers>=4.30.0
huggingface-hub>=0.16.0
torch>=2.0.0
nltk>=3.8.0
spacy>=3.5.0
scikit-learn>=1.0.0
groq

# Data Processing
numpy>=1.21.0

# Visualization
plotly>=5.14.0
matplotlib>=3.5.0

# Additional Dependencies
regex>=2023.0.0
typing-extensions>=4.5.0
requests>=2.31.0
certifi>=2023.0.0